Finding Middle Element of LL
Use the Hare and Tortoise theorem: Take 2 pointers (slow and fast) and initialize both of them to the head of the linked list. Now keep iterating till both the fast pointer and fast->next pointer are NOT NULL and perform the operation of moving the slow pointer by 1 node and fast pointer by 2 nodes. The moment any of the fast or fast->next pointer becomes NULL, the slow pointer would have reached the middle of the linked list. Return the slow pointer.