Create DiameterOfBT.cpp
Diameter is the longest path between any 2 nodes and the path may not be necessarily passing through the root node. That means for the diameter of a tree, if we ae looking for a path passing through a particular node, then for that node, both its left height and right height would be contributing for its diameter because diameter is the path (no of edges) between any 2 nodes.  So  we can find the diameter of a tree by traversing each node once and finding the max of its (maxi, lh + rh) for each node.