Create TopViewofBT.cpp
First take a queue of pair of Node* and int(for line umber) and a map of int and Node*.  First push the root node and corresponding to it line=0 into the queue and pop it and keep pushing the left and right nodes of it  alongside line - 1 and line+1 respectively for the node. nd also keep inserting the {lie and node} into the map. The map should be an ordered one o that it stores the nodes in order. Also the thing to be noted here are that instead of instead of assigning node to line number as a key value pair (like mp[line] = node) , insert the line and node as a pair into the map like mp.insert({line, node}). Now after exiting from the while loop, just extract the value out of the map (it.second->data) into an answer vector.