Create BottomViewinBT.cpp
Same as top view of Binary tree except the fact that instead of inserting line and node into the map as mp.insert({line, node}), we will be directly assigning the node to the line as mp[line] = node. We are doing so because each time we will be encountering any node for any line it will keep updating the value and in the end the bottom node's values would remain.