Create BoundaryTraversal.cpp
First store the left boundary nodes, then the leaf nodes and then the right boundary nodes in reverse order. In order to make sure that we don't store the root node twice while storing the left and right boundary nodes, we won't store the root node at ll. Instead we would store the root node before any traversal, only if the root node is not a leaf node. Then we would store the left boundary nodes starting from root->left nodes ensuring that we are not storing the root node. Then the root nodes and then the right boundary nodes starting from root->right. One thing to note is in order to make sure that we are storing the right boundary nodes in reverse, we would be storing the root->val value after making the recursive call for root->lright and root->left nodes/subtrees.