Create M2.cpp
We traverse in binary tree by in-order method.
We maintain prev node which is just previous to current node in in-order traversal. Initially prev=NULL
First we go to left subtree.
Now for current node there are two possibilities:
If prev==NULL :
      head=root
else :
      prev->right=root
      root->left=prev
Now we make prev=root and go to right subtree.